Reorganize system tests for CAN (#290)
- abstraction layer for system tests defined for easier reuse
- system tests grouped in test classes
- fixed issue with default value of N_Bs timeout always being used
- added a method to CAN Addressing Information class get the other's end Addressing Information
