+++ tests/system_tests/transport_interface/can/common.py
@@ -1,0 +1,19 @@
+from abc import ABC, abstractmethod
+
+from uds.utilities import TimeMillisecondsAlias
+
+
+class AbstractCanTests(ABC):
+    """Abstract class for system tests (with hardware) for Diagnostic over CAN (DoCAN)."""
+
+    TASK_TIMING_TOLERANCE: TimeMillisecondsAlias = 30.  # ms
+    DELAY_AFTER_RECEIVING_FRAME: TimeMillisecondsAlias  # ms
+    DELAY_AFTER_RECEIVING_MESSAGE: TimeMillisecondsAlias  # ms
+
+    @abstractmethod
+    def setup_class(self):
+        """Configure CAN bus connections."""
+
+    @abstractmethod
+    def teardown_class(self):
+        """Safely close CAN bus connections."""
